+++ android/src/main/java/com/reactlibrary/fequency_tools/fft_utils/FastFourierTransform.java
@@ -1,5 +1,5 @@
 package com.reactlibrary.fequency_tools.fft_utils;
 
 public interface FastFourierTransform {
-    float[] calculateSpectrogram(float[] inputData);
+    double[] calculateSpectrogram(double[] inputData);
 }
