+++ android/src/main/java/com/reactlibrary/fequency_tools/windows/HammingWindow.java
@@ -2,15 +2,15 @@ package com.reactlibrary.fequency_tools.windows;
 
 public class HammingWindow implements Window {
     @Override
-    public float[] applyWindow(float[] data) {
-        float[] newData = new float[data.length];
+    public double[] applyWindow(double[] data) {
+        double[] newData = new double[data.length];
         for(int i = 0; i < data.length; i++) {
             newData[i] = data[i] * iterationHamming(i, data.length);
         }
         return newData;
     }
 
-    private float iterationHamming(float n, int size) {
-        return (float)(0.54 - 0.46 * Math.cos(2 * Math.PI * n / size));
+    private double iterationHamming(double n, int size) {
+        return 0.54 - 0.46 * Math.cos(2 * Math.PI * n / (double)size);
     }
 }
